+++ CHANGES.rst
@@ -31,6 +31,7 @@ Development
   for MQTT telemetry to example configurations
 - Fix deep sleep
 - Conditionally start telemetry subsystem just if networking is available
+- Improve robustness wrt. WiFi connectivity
 
 
 2019-06-07 0.3.0